eliminate static cast

+++ source/tessellation/Delaunay.cpp
@@ -1604,7 +1604,7 @@ pair<vector<vector<int> >, vector<int> > Delaunay::FindOuterPoints2
 	{
 		for (size_t j = 0; j < incoming.at(i).size(); ++j)
 		{
-			NghostIndex[i].push_back(static_cast<int>(cor.size() + j));
+			NghostIndex[i].push_back(cor.size() + j);
 			OrgIndex.push_back(cor.size() + j);
 		}
 		AddBoundaryPoints(incoming.at(i));
